# agent_tools/DateTimeAgent/schema/openapi_schema.json
{"openapi": "3.0.3", "info": {"title": "Date and Time Utilities API", "description": "This API provides various date and time utility functions, including current date, time, day, timezone, age calculation, date difference, business day calculation, fiscal year calculation, holiday checks, and SSL/TLS certificate expiration checks.", "version": "1.0.0"}, "servers": [{"url": "/"}], "paths": {"/dt": {"get": {"summary": "GET /dt", "description": "Returns the current date, time, day of the week, and timezone", "operationId": "get_current_date_time_dt_get", "responses": {"422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}, "200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "title": "Return"}}}}}}}, "/age": {"get": {"summary": "GET /age", "description": "Calculates the age based on the given birth date", "operationId": "calculate_age_age_get", "parameters": [{"description": "Birth date in 'YYYY-MM-DD' format", "required": true, "schema": {"type": "string", "title": "Birth Date", "description": "Birth date in 'YYYY-MM-DD' format"}, "name": "birth_date", "in": "query"}], "responses": {"422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}, "200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "title": "Return"}}}}}}}, "/ddiff": {"get": {"summary": "GET /ddiff", "description": "Calculates the difference between two dates in days", "operationId": "calculate_date_diff_ddiff_get", "parameters": [{"description": "Start date in 'YYYY-MM-DD' format", "required": true, "schema": {"type": "string", "title": "Start Date", "description": "Start date in 'YYYY-MM-DD' format"}, "name": "start_date", "in": "query"}, {"description": "End date in 'YYYY-MM-DD' format", "required": true, "schema": {"type": "string", "title": "End Date", "description": "End date in 'YYYY-MM-DD' format"}, "name": "end_date", "in": "query"}], "responses": {"422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}, "200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "title": "Return"}}}}}}}, "/bdays": {"get": {"summary": "GET /bdays", "description": "Calculates the number of business days between two dates", "operationId": "calculate_business_days_bdays_get", "parameters": [{"description": "Start date in 'YYYY-MM-DD' format", "required": true, "schema": {"type": "string", "title": "Start Date", "description": "Start date in 'YYYY-MM-DD' format"}, "name": "start_date", "in": "query"}, {"description": "End date in 'YYYY-MM-DD' format", "required": true, "schema": {"type": "string", "title": "End Date", "description": "End date in 'YYYY-MM-DD' format"}, "name": "end_date", "in": "query"}], "responses": {"422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}, "200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "title": "Return"}}}}}}}, "/fy": {"get": {"summary": "GET /fy", "description": "Calculates the fiscal year for a given date", "operationId": "get_fiscal_year_fy_get", "parameters": [{"description": "Date in 'YYYY-MM-DD' format", "required": true, "schema": {"type": "string", "title": "Date Str", "description": "Date in 'YYYY-MM-DD' format"}, "name": "date_str", "in": "query"}, {"description": "Month in which the fiscal year starts (default is 10 for October)", "required": false, "schema": {"type": "integer", "title": "Fiscal Year Start Month", "description": "Month in which the fiscal year starts (default is 10 for October)", "default": 10}, "name": "fiscal_year_start_month", "in": "query"}], "responses": {"422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}, "200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "title": "Return"}}}}}}}, "/fyr": {"get": {"summary": "GET /fyr", "description": "Calculates the start and end dates of the fiscal year for a given date", "operationId": "get_fiscal_year_range_fyr_get", "parameters": [{"description": "Date in 'YYYY-MM-DD' format", "required": true, "schema": {"type": "string", "title": "Date Str", "description": "Date in 'YYYY-MM-DD' format"}, "name": "date_str", "in": "query"}, {"description": "Month in which the fiscal year starts (default is 10 for October)", "required": false, "schema": {"type": "integer", "title": "Fiscal Year Start Month", "description": "Month in which the fiscal year starts (default is 10 for October)", "default": 10}, "name": "fiscal_year_start_month", "in": "query"}], "responses": {"422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}, "200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "title": "Return"}}}}}}}, "/nbday": {"get": {"summary": "GET /nbday", "description": "Calculates the next business day after the given date", "operationId": "get_next_business_day_nbday_get", "parameters": [{"description": "Date in 'YYYY-MM-DD' format", "required": true, "schema": {"type": "string", "title": "Date Str", "description": "Date in 'YYYY-MM-DD' format"}, "name": "date_str", "in": "query"}], "responses": {"422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}, "200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "title": "Return"}}}}}}}, "/pdl": {"get": {"summary": "GET /pdl", "description": "Checks if a policy has expired based on the given expiration date", "operationId": "check_policy_deadline_pdl_get", "parameters": [{"description": "Name of the policy", "required": true, "schema": {"type": "string", "title": "Policy Name", "description": "Name of the policy"}, "name": "policy_name", "in": "query"}, {"description": "Expiration date in 'YYYY-MM-DD HH:MM:SS' format", "required": true, "schema": {"type": "string", "title": "Expiry Date", "description": "Expiration date in 'YYYY-MM-DD HH:MM:SS' format"}, "name": "expiry_date", "in": "query"}], "responses": {"422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}, "200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "title": "Return"}}}}}}}, "/ssl": {"get": {"summary": "GET /ssl", "description": "Checks if an SSL/TLS certificate has expired for a given hostname", "operationId": "check_ssl_cert_deadline_ssl_get", "parameters": [{"description": "Hostname for which to check the SSL/TLS certificate", "required": true, "schema": {"type": "string", "title": "Hostname", "description": "Hostname for which to check the SSL/TLS certificate"}, "name": "hostname", "in": "query"}, {"description": "Port number for the SSL/TLS connection (default is 443 for HTTPS)", "required": false, "schema": {"type": "integer", "title": "Port", "description": "Port number for the SSL/TLS connection (default is 443 for HTTPS)", "default": 443}, "name": "port", "in": "query"}], "responses": {"422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}, "200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "title": "Return"}}}}}}}}, "components": {"schemas": {"HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}, "tags": [{"name": "date"}, {"name": "time"}, {"name": "utilities"}]}